[dev] Move all dev scripts into shared dir.
